Backend/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
requests==2.31.0
pydantic==2.3.0
python-dotenv==1.0.0
pandas==2.1.0
langchain==0.0.312
pinecone-client==2.2.4
python-multipart==0.0.6